Add files via upload
A simple command line text based game I built to experiment with a range of variables, while loops, if and else if statements, and the Scanner class that is used for getting the input of different primitive data types. 

note: file must be compiled into bytecode class files in terminal using the command javac.